refactor: inline execution-only helper modules #277
